+++ app/Http/Controllers/User/TicketController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 use App\Models\CustomUserTable;
+use App\Models\UserProfile;
 use App\Models\Ticket;
 use App\Models\Office;
 
@@ -107,6 +108,78 @@ class TicketController extends Controller
                     ], 404);
                 }
             }
+        else 
+        {
+            // If the User is Anonymous
+            return response()->json([
+                'status' => 'error',
+                'message' => "Unauthorized Access.",
+            ], 409);
+        }
+    }
+
+    // Get Ticket Info
+    public function backend_getTicketInfo(Request $request, $ticket_number)
+    {
+        if (Auth::check()) {
+            // Check if the user is an admin
+            if (Auth::user()->is_admin) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => "Unauthorized Access.",
+                ], 409);
+            } else {
+                $ticket = Ticket::where('ticket_number', $ticket_number)->first();
+
+                $ticketdata = [];
+
+                if ($ticket) {
+
+                    $office = Office::where('office_id', $ticket->office_id)->first();
+                    $user = UserProfile::where('user_id', $ticket->user_id)->first();
+
+                    $ticketData[] = [
+                        'ticket_id' => $ticket->ticket_id,
+                        'affiliation' => $ticket->affiliation,
+                        'created_at' => $ticket->created_at,
+                        'ticket_number' => $ticket->ticket_number,
+                        'priority' => $ticket->priority,
+                        'status' => $ticket->status,
+                        'type' => $ticket->type,
+                        'office' => $office->office_name,
+                        'service' => $ticket->service,
+                        'resolved_date' => $ticket->resolved_date,
+                        'user' => $user->last_name." ".$user->first_name,
+                    ];
+
+                    // user Verification
+                    $user = Auth::user();
+
+                    if ($user->user_id == $ticket->user_id) {
+                        return response()->json([
+                            'status' => 'success',
+                            'message' => "Access Granted.",
+                            'data' => $ticketData,
+                            'verify' => 'matched',
+                        ], 200);
+                    }
+                    else {
+                        return response()->json([
+                            'status' => 'error',
+                            'verify' => 'unmatched',
+                            'message' => "Access Denied.",
+                        ], 200);
+                    }
+
+                }
+                else {
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => "Data not found.",
+                    ], 404);
+                }
+            }
+        }
         else 
         {
             // If the User is Anonymous